Move mxcp.core -> sdk
