+++ pyproject.toml
@@ -66,7 +66,7 @@ all = [
 
 [tool.setuptools]
 package-dir = {"" = "src"}
-packages = ["mxcp", "mxcp.audit", "mxcp.cli", "mxcp.config", "mxcp.config.schemas", "mxcp.server", "mxcp.auth", "mxcp.drift", "mxcp.drift.schemas", "mxcp.endpoints", "mxcp.endpoints.schemas", "mxcp.engine", "mxcp.plugins", "mxcp.policies", "mxcp.evals", "mxcp.evals.schemas", "mxcp.runtime", "mxcp.core", "mxcp.core.analytics", "mxcp.core.config", "mxcp.core.config.resolvers", "mxcp.core.config.schemas", "mxcp.validator", "mxcp.validator.schemas"]
+packages = ["mxcp", "mxcp.cli", "mxcp.config", "mxcp.config.schemas", "mxcp.dbt", "mxcp.drift.schemas", "mxcp.endpoints.schemas", "mxcp.evals", "mxcp.evals.schemas", "mxcp.plugins", "mxcp.policies", "mxcp.runtime", "mxcp.sdk", "mxcp.sdk.audit", "mxcp.sdk.auth", "mxcp.sdk.core", "mxcp.sdk.core.analytics", "mxcp.sdk.core.config", "mxcp.sdk.core.config.resolvers", "mxcp.sdk.evals", "mxcp.sdk.executor", "mxcp.sdk.executor.plugins", "mxcp.sdk.executor.plugins.duckdb_plugin", "mxcp.sdk.executor.plugins.python_plugin", "mxcp.sdk.policy", "mxcp.sdk.validator", "mxcp.validator"]
 include-package-data = true
 
 [tool.setuptools.package-data]
@@ -74,7 +74,7 @@ include-package-data = true
 "mxcp.endpoints.schemas" = ["*.json"]
 "mxcp.drift.schemas" = ["*.json"]
 "mxcp.evals.schemas" = ["*.json"]
-"mxcp.core.config.schemas" = ["*.json"]
+"mxcp.sdk.core.config.schemas" = ["*.json"]
 "mxcp.validator.schemas" = ["*.json"]
 
 [tool.pytest.ini_options]
