+++ src/mxcp/sdk/core/auth_helpers.py
