+++ src/mxcp/sdk/core/config/schemas/resolver-config-schema.json
